requirements.txt
streamlit==1.28.0
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
requests==2.31.0